add email authenticastion

+++ App/src/app/components/register/register.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { AuthService } from 'src/app/services/auth.service';
 import { FormBuilder,FormControl,FormGroup,Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
+
 
 @Component({
   selector: 'app-register',
@@ -10,6 +12,7 @@ import { Router } from '@angular/router';
 })
 export class RegisterComponent implements OnInit {
   confpassword:string;
+  imageHidden:boolean=true;
   sub:boolean=false
   registerUserData={
     email:'',
@@ -20,22 +23,39 @@ export class RegisterComponent implements OnInit {
   }
   constructor(
     public _auth:AuthService,
-    public router:Router
+    public router:Router,
+    public toastr:ToastrService
   ) { }
 
   ngOnInit(): void {
   }
 
+  showSuccess() {
+   this.toastr.warning("User is Exist", "Registration Failed");
+  }
+
   registerUser(){
+    this.imageHidden=false;
     this._auth.registerUser(this.registerUserData).subscribe(
       res=>{
-        console.log(res)
-        localStorage.setItem('token', res.token)
-        this.router.navigate(['/special'])
+
+        if(res.exist=="yes"){
+          console.log("Available")
+          this.showSuccess()
+         this.imageHidden=true;
+        }
+
+        else{
+          console.log("Register Success")
+          localStorage.setItem('token', res.token)
+          this.router.navigate(['/login'])
+        // this.imageHidden=true;
+        }
 
       },
       err=>{
         console.log(err)
+
       }
     )
   }
